Split LicenseKnowledgeBaseResolver in processor and implementation
